src/app/requirements.txt
fastapi==0.95.0
google-cloud-bigquery==2.34.4
numpy==1.23.5
pandas==1.5.3
uvicorn==0.21.1